+++ hutool-core/src/main/java/cn/hutool/core/text/UnicodeUtil.java
@@ -1,5 +1,6 @@
 package cn.hutool.core.text;
 
+import cn.hutool.core.lang.Console;
 import cn.hutool.core.util.CharUtil;
 import cn.hutool.core.util.HexUtil;
 import cn.hutool.core.util.StrUtil;